refactor(mcp): split handlers into submodules
Split large handlers.rs file into module directory structure:
- mod.rs: Public API and exports
- batch.rs: Batch operation handlers
- tool_call.rs: Tool call dispatch and handlers

File size compliance: Each file now <500 LOC.
